PycharmProjects/python_learning/Kafka/kafka_admin_api_start/requirements.txt
kafka-python==2.3.0
fastapi==0.128.0
python-dotenv==1.2.1
uvicorn==0.22.0
python-dotenv==1.2.1

